requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
python-dotenv==1.1.1
flask-cors==6.0.1